+++ app/src/main/java/io/fastpix/app/ui/main/VideoPlayerActivity.kt
@@ -15,6 +15,7 @@ import android.os.Looper
 import android.util.Log
 import android.view.View
 import android.view.WindowInsetsController
+import android.view.WindowManager
 import android.widget.SeekBar
 import androidx.activity.OnBackPressedCallback
 import androidx.annotation.OptIn
@@ -69,6 +70,7 @@ class VideoPlayerActivity : AppCompatActivity() {
         super.onCreate(savedInstanceState)
         binding = ActivityVideoPlayerBinding.inflate(layoutInflater)
         setContentView(binding.root)
+        window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
             videoModel = intent.getParcelableExtra("video_model", DummyData::class.java)
         } else {
@@ -444,9 +446,16 @@ class VideoPlayerActivity : AppCompatActivity() {
 
     private fun formatTime(timeMs: Long): String {
         val totalSeconds = timeMs / 1000
-        val minutes = totalSeconds / 60
+
+        val hours = totalSeconds / 3600
+        val minutes = (totalSeconds % 3600) / 60
         val seconds = totalSeconds % 60
-        return String.format(Locale.getDefault(), "%02d:%02d", minutes, seconds)
+
+        return if (hours > 0) {
+            String.format(Locale.getDefault(), "%02d:%02d:%02d", hours, minutes, seconds)
+        } else {
+            String.format(Locale.getDefault(), "%02d:%02d", minutes, seconds)
+        }
     }
 
     override fun onPause() {